Rename handlers to workers
