+++ examples/ConductorSharp.NoApi/Program.cs
@@ -1,8 +1,7 @@
 ﻿using ConductorSharp.Engine.Extensions;
 using ConductorSharp.Engine.Health;
-using ConductorSharp.KafkaCancellationNotifier.Extensions;
-using ConductorSharp.NoApi.Handlers;
 using ConductorSharp.NoApi.Middlewares;
+using ConductorSharp.NoApi.Workers;
 using ConductorSharp.Patterns.Extensions;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -43,9 +42,9 @@ var builder = Host.CreateDefaultBuilder()
             //    topicName: configuration.GetValue<string>("Conductor:KafkaCancellationNotifier:TopicName"),
             //    groupId: configuration.GetValue<string>("Conductor:KafkaCancellationNotifier:GroupId"));
 
-            services.RegisterWorkerTask<GetCustomerHandler>();
-            services.RegisterWorkerTask<PrepareEmailHandler>();
-            services.RegisterWorkerTask<EnumTaskHandler>();
+            services.RegisterWorkerTask<GetCustomerWorker>();
+            services.RegisterWorkerTask<PrepareEmailWorker>();
+            services.RegisterWorkerTask<EnumTaskWorker>();
         }
     );
